Change Reverse Engineering ForeignKeyProcessor HashMap to TreeMap
The order of One-To-Many relations in POJOs was not explicitly defined.
The Properties were inserted into a HashMap, using the FK name as key.
Depending on the capacity and load factor of the HashMap, the ordering
of the relations changed when the number of foreign keys in a given
class changed and resulted in a different actual capacity of the HashMap.
By using a TreeMap, the relations are always sorted by their respective
FK's name which is consistent to the ordering of the To-One-Relations.